changing all id to long id + fixing osmapi fetcher to match seamsless cities + using newton json parser

+++ LittleConqueror.AppService/Domain/DrivingModels/Commands/AddCityToATerritoryCommand.cs
@@ -2,7 +2,7 @@ namespace LittleConqueror.AppService.Domain.DrivingModels.Commands;
 
 public class AddCityToATerritoryCommand
 {
-    public int CityId { get; set; }
+    public long CityId { get; set; }
     public char CityType { get; set; }
-    public int TerritoryId { get; set; }
+    public long TerritoryId { get; set; }
 }
