+++ LittleConqueror.AppService/Domain/DrivingModels/Commands/CreateUserCommand.cs
@@ -3,6 +3,6 @@ namespace LittleConqueror.AppService.Domain.DrivingModels.Commands;
 public class CreateUserCommand
 {
     public string Name { get; set; }
-    public int FirstOsmId { get; set; }
+    public long FirstOsmId { get; set; }
     public char FirstOsmType { get; set; }
 }
