+++ LittleConqueror.AppService/Domain/DrivingModels/Queries/GetCityByOsmIdQuery.cs
@@ -2,6 +2,6 @@ namespace LittleConqueror.AppService.Domain.DrivingModels.Queries;
 
 public class GetCityByOsmIdQuery
 {
-    public int OsmId { get; set; }
+    public long OsmId { get; set; }
     public char OsmType { get; set; }
 }
