+++ LittleConqueror.AppService/DrivenPorts/ICityDatabasePort.cs
@@ -4,7 +4,7 @@ namespace LittleConqueror.AppService.DrivenPorts;
 
 public interface ICityDatabasePort
 {
-    Task<City?> GetCityById(int id);
+    Task<City?> GetCityById(long id);
     Task<City?> AddCity(City city);
-    Task SetTerritoryId(int cityId, int territoryId);
+    Task SetTerritoryId(long cityId, long territoryId);
 }
