+++ LittleConqueror.AppService/DrivenPorts/IResourcesDatabasePort.cs
@@ -5,5 +5,5 @@ namespace LittleConqueror.AppService.DrivenPorts;
 public interface IResourcesDatabasePort
 {
     public Task<Resources> CreateResources(Resources resources);
-    public Task<Resources?> GetResourcesOfUser(int userId);
+    public Task<Resources?> GetResourcesOfUser(long userId);
 }
