+++ LittleConqueror.AppService/DrivenPorts/ITerritoryDatabasePort.cs
@@ -5,6 +5,6 @@ namespace LittleConqueror.AppService.DrivenPorts;
 public interface ITerritoryDatabasePort
 {
     public Task<Territory> CreateTerritory(Territory territory);
-    public Task<Territory?> GetTerritoryOfUser(int userId);
-    public Task<Territory?> GetTerritoryById(int territoryId);
+    public Task<Territory?> GetTerritoryOfUser(long userId);
+    public Task<Territory?> GetTerritoryById(long territoryId);
 }
