+++ LittleConqueror.Infrastructure/DatabaseAdapters/CityDatabaseAdapter.cs
@@ -8,7 +8,7 @@ namespace LittleConqueror.Infrastructure.DatabaseAdapters;
 public class CityDatabaseAdapter(
     CityRepository cityRepository) : ICityDatabasePort
 {
-    public async Task<City?> GetCityById(int id)
+    public async Task<City?> GetCityById(long id)
         => await cityRepository.GetAsync(cityEntity => cityEntity.Id == id);
 
     public async Task<City?> AddCity(City city)
@@ -20,7 +20,7 @@ public class CityDatabaseAdapter(
         return entityEntry.Entity;
     }
 
-    public async Task SetTerritoryId(int cityId, int territoryId)
+    public async Task SetTerritoryId(long cityId, long territoryId)
     {
         await ValidateIfNotExistAsync(cityId);
         
@@ -33,7 +33,7 @@ public class CityDatabaseAdapter(
         await cityRepository.UpdateAsync(city);
     }
     
-    private async Task ValidateIfNotExistAsync(int Id)
+    private async Task ValidateIfNotExistAsync(long Id)
     {
         var existingEntity = await cityRepository.GetAsync(entity => entity.Id == Id);
         if (existingEntity == null)
