+++ LittleConqueror/API/Models/Dtos/TerritoryDto.cs
@@ -2,8 +2,8 @@ namespace LittleConqueror.API.Models.Dtos;
 
 public class TerritoryDto
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     
-    public int OwnerId { get; set; }
+    public long OwnerId { get; set; }
     public IEnumerable<CityDto> Cities { get; set; }
 }
